Fix issue with Python 3
In Python 3 all divisions between int values result in a float which
can't be used as index. Therefore adjust the code to cast the calculation
result to an int.

+++ weather.py
@@ -111,8 +111,8 @@ class Sensor(db.Model):
         if length == 0:
             return 0
         if not length % 2:
-            return (s[length / 2] + s[length / 2] - 1) / 2
-        return s[length / 2]
+            return (s[int(length / 2)] + s[int(length / 2)] - 1) / 2
+        return s[int(length / 2)]
 
     @staticmethod
     def _round_to_hour(dt):
